hybrid_array::sizes

Module extra_sizes

Additional typenum size aliases beyond what are normally provided.

These are defined using their component bits rather than Add to avoid conflicting impls.

Macros

uint 🔒

Type Aliases

U749

U781

U877

U897

U994

U1039

U1040

U1056

U1072

U1088

U1104

U1120

U1124

U1136

U1152

U1153

U1156

U1158

U1168

U1184

U1200

U1216

U1232

U1248

U1249

U1252

U1264

U1280

U1296

U1312

U1322

U1328

U1344

U1349

U1360

U1376

U1392

U1408

U1424

U1440

U1455

U1456

U1472

U1488

U1498

U1504

U1505

U1518

U1520

U1530

U1536

U1552

U1568

U1584

U1600

U1616

U1623

U1632

U1648

U1658

U1664

U1665

U1680

U1696

U1712

U1728

U1744

U1760

U1763

U1776

U1792

U1808

U1824

U1840

U1847

U1856

U1872

U1888

U1904

U1920

U1936

U1952

U1968

U1984

U1999

U2000

U2016

U2032

U2047

U2064

U2067

U2080

U2096

U2112

U2128

U2144

U2160

U2176

U2180

U2192

U2208

U2224

U2240

U2241

U2254

U2256

U2272

U2288

U2304

U2320

U2336

U2352

U2368

U2384

U2400

U2416

U2417

U2420

U2432

U2448

U2464

U2480

U2496

U2512

U2528

U2544

U2560

U2576

U2592

U2608

U2624

U2640

U2656

U2672

U2688

U2704

U2720

U2736

U2752

U2768

U2784

U2800

U2816

U2832

U2848

U2864

U2880

U2896

U2912

U2928

U2944

U2960

U2976

U2992

U3008

U3024

U3040

U3056

U3059

U3072

U3088

U3104

U3120

U3136

U3152

U3168

U3184

U3200

U3216

U3232

U3248

U3264

U3280

U3296

U3309

U3312

U3328

U3344

U3360

U3376

U3392

U3408

U3424

U3440

U3456

U3472

U3488

U3504

U3520

U3536

U3552

U3568

U3584

U3600

U3616

U3632

U3648

U3664

U3680

U3696

U3712

U3728

U3744

U3760

U3776

U3792

U3808

U3824

U3840

U3856

U3872

U3888

U3904

U3920

U3936

U3952

U3968

U3984

U4000

U4016

U4032

U4048

U4064

U4080

U4292

U4433

U4480

U4514

U4544

U4595

U4627

U4896

U6492

U7237

U7856

U8160

U8516

U8978

U9616

U9720

U9752

U12240

U13608

U13932

U13948

U14120

U14421

U15632

U15744

U15792

U16224

U16320

U17088

U19888

U21520

U21632

U21696

U29792

U31296

U35664

U43088

U49856

U261120

U524160

U1044992

U1047319

U1357824